refactor: move getDistributions to bounds utils
